Integrate the HomepageFeatures component directly into index.js
Integrate the HomepageFeatures component directly into index.js
